pyproject.toml
[project]
name = "api-bancaria"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "aiosqlite>=0.21.0",
    "alembic>=1.17.2",
    "fastapi[standard]>=0.121.2",
    "pyjwt>=2.10.1",
    "python-dotenv>=1.2.1",
    "sqlalchemy>=2.0.44",
]
